requirements.txt
# Updated requirements.txt with correct versions and dependencies

# Core Framework
crewai==0.130.0
crewai-tools==0.47.1

# FastAPI and Web Framework
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.6

# LangChain and AI
langchain==0.3.27
langchain-openai==0.2.14
langchain-community==0.3.28
langchain-core==0.3.75
openai==1.102.0

# PDF Processing
pdfplumber==0.11.7
pypdf==5.9.0
pdfminer.six==20250506

# Environment and Configuration
python-dotenv==1.1.1
pydantic==2.11.7
pydantic-settings==2.10.1

# Data Processing
pandas==2.3.2
numpy==2.3.2

# HTTP and Networking
requests==2.32.5
httpx==0.28.1
aiohttp==3.12.15

# Utilities
typing-extensions==4.15.0
click==8.1.8
rich==13.9.4
tenacity==9.1.2

# Development and Testing
pytest==8.3.2
pytest-asyncio==0.24.0

# Optional: Database support (if needed later)
sqlalchemy==2.0.43

# Optional: Caching (if needed)
redis==5.2.0

# Optional: Monitoring and Logging
structlog==24.4.0